pdf_processor/requirements.txt
# PDF文档智能处理系统依赖项

# 基础依赖
numpy>=1.24.0
pandas>=1.5.0

# PDF处理相关
PyPDF2>=3.0.0
pdf2image>=1.16.3

# OCR相关
paddleocr>=2.7.0
paddlepaddle>=2.5.0

# 图像处理
Pillow>=9.5.0
opencv-python>=4.7.0

# 文本处理
markdown>=3.4.0
beautifulsoup4>=4.12.0

# 工具库
tqdm>=4.65.0

# 开发工具
pytest>=7.3.0
black>=23.3.0
flake8>=6.0.0
